% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwqr-package.R
\docType{package}
\name{dwqr-package}
\alias{dwqr}
\alias{dwqr-package}
\title{dwqr: **D**rinking **W**ater **Q**uality **R**}
\description{
This package allows for calculation of different statistics 
    for drinking water quality in drinking water distribution systems.
}
\keyword{internal}
