% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dwqr_datasets}
\alias{dwqr_datasets}
\alias{nitrification}
\title{\strong{dwqr datasets}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 536 rows and 4 columns.}
\usage{
nitrification
}
\description{
An example dataset in the correct format for use with the \code{dwqr}
nitrification functions (e.g nitrification_al, plot_fl, plot_al)
}
\keyword{datasets}
