% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falling-limb.R
\name{falling_limb}
\alias{falling_limb}
\title{classify the falling limb of chlorine trend}
\usage{
falling_limb(
  data,
  method = c("simple", "hmm", "cp"),
  value_col,
  first_deriv,
  group_col
)
}
\arguments{
\item{data}{a data frame with chlorine residual results}

\item{method}{the method used to classify the falling limb of the chlorine
curve. See Details.}

\item{value_col}{unqouted column name of column containing the chlorine
results for the time series}

\item{first_deriv}{unquoted column name of column containing first derivative
of chlorine time series}

\item{group_col}{vector of unqouted column names of any grouping columns}
}
\description{
\code{falling_limb} use the first and second derivative of a
time series slope to classify portions of it as the falling limb
of the curve
}
\details{
the method argument must be set to one of the following:
\itemize{
\item "simple" - A simple classification method that classifies any negative
first derivative value as a part of the falling limb. Taking the first
derivative of the moving average of the chlorine values is likely to reduce
false classification rates when this model type is selected
\item "hmm" - This method uses the depmixS4 package to fit a hidden markov model
using the time trend of the first derivative of the total chlorine trend
\item "cp" - This method uses the strucchange package to identify change points
in the first derivative trned and classify values based on median first derivative
values between changepoints
}
}
