% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrification-al.R
\name{nitrification_al}
\alias{nitrification_al}
\title{determine action levels for nitrification}
\usage{
nitrification_al(data, date_col, value_col, ..., method = c("FL", "P"),
  percentiles = c(0.8, 0.5, 0.1), rolling_window = 8,
  max_chlorine = 1.5, output_name = c("AL-C", "AL", "P"))
}
\arguments{
\item{data}{a data frame with chlorine residual results}

\item{date_col}{the unquoted column name of a date or datetime column in data}

\item{value_col}{the unquoted column name of the results in data}

\item{...}{unquoted column names of grouping variables}

\item{method}{either "FL" (Falling Limb) or "P" (Percentiles)}

\item{percentiles}{a vector of the percentiles that you would like calculated
either on the overall data ("P" method) or on the falling limb portion of
the dataset ("FL" method). By default, 0.8, 0.5, and 0.1 are used for action
levels 1, 2, and 3, respectively}

\item{rolling_window}{how many observations should be included in the rolling
average window function when calculating the first and second derivative of
the chlorine time series. Defaults to 8.}

\item{max_chlorine}{maximum chlorine residual value that can be included in
falling limb. For example, if you are not concerned with sites when chlorine
is greater than 1.5 (default) than no value greater than this threshold
will be classified as either "Falling Limb" or "Nitrification Ongoing"}

\item{output_name}{should the output column names be given as title
action levels ("AL-C"), i.e. Action Level 1, Action Level 2, etc.,
as action levels for R code ("AL"), i.e. action_level_1, etc.,
or as the percentile ("P"), i.e.
80%, 50%, etc.}
}
\description{
\code{nitrification_al} calculates proposed action levels for chlorine
residual for nitrification control and response plans. These action
levels can be calculated via the falling limb method or by plain
percentiles of the overall data
}
\details{
There are two methods provided for calculating nitrification action levels:

FL - \emph{Falling Limb} - This method uses the falling limb method to identify
the downward trend experienced by chlorine values as temperatures rise. Before
percentiles are caluclated, the falling limb portion of the graph is isolated
and only this portion is used in the percentile calculations

P - \emph{Percentiles} - This method simply takes the percentiles given on the
entire dataset to find the action levels.

By default, there are three percentiles calculated (80%, 50%, 10%) corresponding
to three distinct action levels (80% - Action Level 1, 50% - Action Level 2,
10% - Action Level 3). If more or less action levels are desired, then simply
specify more or less values in the \strong{percentiles} argument.

The output tibble will contain columns for each of the specified columns along
with any grouping variables used. By default, the column names will be given as
their corresponding action level. If you would like to display the percentiles as
column names rather than action levels, then specify \strong{output_name = "P"}

Both of these methods will struggle if samples are not collected regularly at
a given location. Regular sampling allows for the overall trends to be identified
and classified correctly.
}
