% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-al.R
\name{plot_al}
\alias{plot_al}
\title{Plot the action levels of chlorine results}
\usage{
plot_al(data, date_col, value_col, ..., method = c("FL", "P"),
  percentiles = c(0.8, 0.5, 0.1), rolling_window = 8,
  max_chlorine = 1.5, date_breaks = "6 months",
  date_labels = "\%b \%d, \%Y", ylab = "", plot_title = "",
  plot_subtitle = "", legend_title = "", action_levels = NULL)
}
\arguments{
\item{data}{a data frame or tibble that has chlorine residual data}

\item{date_col}{the unquoted column name of the date/datetime column}

\item{value_col}{the unqouted column name of the chlorine residual column}

\item{...}{unqouted column names of all grouping columns}

\item{method}{either "FL" or "P" to distinguish how you want the
action levels to be calculated}

\item{percentiles}{the percentiles that will be used to calculate the
action levels of either the falling limb (if method = "FL") or the
overall distribution (if method = "P")}

\item{rolling_window}{how many observations should be included in the rolling
average window function when calculating the first and second derivative of
the chlorine time series. Defaults to 8.}

\item{max_chlorine}{maximum chlorine residual value that can be included in
falling limb. For example, if you are not concerned with sites when chlorine
is greater than 1.5 (default) than no value greater than this threshold
will be classified as either "Falling Limb" or "Nitrification Ongoing"}

\item{date_breaks}{a character string to be passed to either
\code{\link{ggplot2::scale_x_date}} or \code{\link{ggplot2::scale_x_datetime}}
to specify the spacing of date breaks (e.g. "1 month" or "3 weeks")}

\item{date_labels}{a character string specifying the desired output
format of dates on x axis. See \code{\link{strptime}} for options}

\item{ylab}{a character string specifying the y axis label for the main plot}

\item{plot_title}{a character string specifying the title of the plot}

\item{plot_subtitle}{a character string specifying the subtitle of the plot}

\item{legend_title}{a character string specifying the legend title}

\item{action_levels}{a numeric vector specifying action levels. If this is
NULL (default) than the action levels will be specified via the method
provided in the method arguement. If this argument is given, then the vector
will be sorted with the lowest number by default being the highest action level
and the highest number given being the lowest action level. The number of action
levels in the resulting data will be equal to the number of action levels given}
}
\description{
Generate plots showing whether an action level is exceeded
}
