% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fl.R
\name{plot_fl}
\alias{plot_fl}
\title{Plot falling limb functions}
\usage{
plot_fl(data, date_col, value_col, ..., rolling_window = 8,
  max_chlorine = 1.5, date_breaks = "6 months",
  date_labels = "\%b \%d, \%Y", ylab = "", plot_title = "",
  plot_subtitle = "", include_first = FALSE, theme = NULL,
  nitrite_col, nitrite_ylab = "Nitrite")
}
\arguments{
\item{data}{a data frame containing chlorine residual results}

\item{date_col}{unqouted column name of date time column in data}

\item{value_col}{unquoted column name of chlorine column in data}

\item{...}{unqouted column name(s) of grouping variable(s)}

\item{rolling_window}{how many observations should be included in the rolling
average window function when calculating the first and second derivative of
the chlorine time series. Defaults to 8.}

\item{max_chlorine}{maximum chlorine residual value that can be included in
falling limb. For example, if you are not concerned with sites when chlorine
is greater than 1.5 (default) than no value greater than this threshold
will be classified as either "Falling Limb" or "Nitrification Ongoing"}

\item{date_breaks}{a character string to be passed to either
\code{\link{ggplot2::scale_x_date}} or \code{\link{ggplot2::scale_x_datetime}}
to specify the spacing of date breaks (e.g. "1 month" or "3 weeks")}

\item{date_labels}{a character string specifying the desired output
format of dates on x axis. See \code{\link{strptime}} for options}

\item{ylab}{a character string specifying the y axis label for the main plot}

\item{plot_title}{a character string specifying the title of the plot}

\item{plot_subtitle}{a chacter string specifying the subtitle of the plot}

\item{include_first}{logical; should a plot of the first derivative and
its moving average be included along with the main trend?}

\item{theme}{a \code{\link{ggplot2::theme}} object that can be passed to
all plots}

\item{nitrite_col}{column name (optional) of column containing nitrite
data that corresponds with the chlorine data. This will be plotted
alongside the chlorine data via the \code{\link{patchwork}} package}

\item{nitrite_ylab}{character string that specifies the y label for the
nitrite graph. Only used if nitrite_col is specified}
}
\description{
\code{plot_fl} provides methods to plot the falling limb of a
chlorine residual time series plot. This function will be useful for
trying to visualize how selected parameters such as \code{rolling_window}
effect the grouping of data points
}
