% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falling-limb.R
\name{falling_limb}
\alias{falling_limb}
\title{classify the falling limb of chlorine trend}
\usage{
falling_limb(data, value_col, first_deriv, second_deriv, ...,
  max_chlorine = 1.5)
}
\arguments{
\item{data}{a data frame with chlorine residual results}

\item{value_col}{unqouted column name of column containing the chlorine
results for the time series}

\item{first_deriv}{unquoted column name of column containing first derivative
of chlorine time series}

\item{second_deriv}{unquoted column name of column containing second
derivative of chlorine time series}

\item{...}{unqouted column names of any grouping columns}

\item{max_chlorine}{maximum chlorine residual value that can be included in
falling limb. For example, if you are not concerned with sites when chlorine
is greater than 1.5 (default) than no value greater than this threshold
will be classified as either "Falling Limb" or "Nitrification Ongoing"}
}
\description{
\code{falling_limb} use the first and second derivative of a
time series slope to classify portions of it as the falling limb
of the curve
}
