% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling-slope.R
\name{rolling_slope}
\alias{rolling_slope}
\title{Calculate first and second derivative of chlorine trend}
\usage{
rolling_slope(data, date_col, value_col, ..., rolling_window = 8)
}
\arguments{
\item{data}{a dataframe or tbl that contains at the minimum a date column
and a chlorine residual column}

\item{date_col}{the unqouted name of the date column}

\item{value_col}{the unquoted name of the chlorine residual column}

\item{...}{the unqouted names of all grouping columns, if grouping is
needed - e.g. if the dataset contains multiple sites and/or parameters.}

\item{rolling_window}{how large should the rolling mean window be?}
}
\description{
\code{rolling_slope} calculates the first and second derivative
of a provided time series trend along with the moving average
of these values
}
